requirements.txt
python-telegram-bot
chromadb
openai
python-dotenv
langchain
langchain-openai
tiktoken